Add SPIR-V-related tools. (#8)
This adds two projects related to the SPIR-V format used in APIs such as OpenCL and Level Zero:

* https://github.com/KhronosGroup/SPIRV-LLVM-Translator: Bidirectional translator between SPIR-V and LLVM IR.
* https://github.com/KhronosGroup/SPIRV-Tools: Assortment of tools surrounding SPIR-V (assembler, disassembler, optimizer).

The projects depend on our LLVM build so their versions need to be kept compatible in case we update LLVM. They are built out-of-tree.

The projects are excluded from the CentOS build because we're not able to build them correctly on CentOS (at least the git is too outdated, probably the C++ compiler as well). I think this is acceptable because we likely won't use anything related to SPIR-V on CentOS machines.