Build LLVM with static libraries (#12)
* Build LLVM with static libs.

This causes problems with our existing OpenCL implementation, which
requires the static libraries.  Excerpt from its cmake config:

```
Explicitly link against the individual static versions of the clang
libraries, as if clang was built with LLVM_LINK_LLVM_DYLIB=ON, then the
'clangCodeGen' library target brings in the 'LLVM' dynamic library. It is
not safe for us to link with LLVM's dynamic library, due to LLVM's use of
global/static state.
```

Also, the documentation of `BUILD_SHARED_LIBS` in the LLVM docs state:

```
BUILD_SHARED_LIBS is only recommended for use by LLVM developers. If you
want to build LLVM as a shared library, you should use the
LLVM_BUILD_LLVM_DYLIB option.
```

Hence, I think we should just leave this with the default.  Our
customers aren't LLVM developers.

* Remove explicit LLVM cmake option.

Just leave this with the default.  It speeds up creating Debug builds of
LLVM, but we're doing a Release build anyway.